Design: 검색창 디자인 수정 및 drawer 와 dialog 설정
