+++ apps/client/src/components/common/organism/SearchOrganism.tsx
@@ -3,7 +3,7 @@
 import React, { useState, useEffect } from "react"
 import debounce from "lodash/debounce"
 import { useSearchActions } from "@/action/search/useSearchResults"
-import SearchInput from "../atom/SearchInput"
+import SearchInput from "../atom/SearchInputWrapper"
 import SearchDropdown from "../molecule/SearchDropdown"
 import { useRouter } from "next/navigation"
